% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_snapshot.R
\name{compare_snapshot}
\alias{compare_snapshot}
\title{Compare your session info with packages stored in `config/packages.dcf`}
\usage{
compare_snapshot(only_attached = T)
}
\arguments{
\item{only_attached}{if TRUE, it will only compare attached packages.}
}
\description{
A function which allows you to compare your set of packages with those stored in `config/packages.dcf` file.
}
\examples{
\dontrun{
 compare_snapshot()
 }
}
