% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_snapshot.R
\name{docker_snapshot}
\alias{docker_snapshot}
\title{Create a chunk of code with package installation in Dockerfile}
\usage{
docker_snapshot()
}
\description{
Create a chunk of code with package installation in Dockerfile
}
\examples{
\dontrun{
 snapshot_pkg()
 }
}
