% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_create.R
\name{project_create}
\alias{project_create}
\title{Create a new ENR project.}
\usage{
project_create(project_name = "enr_project", path)
}
\arguments{
\item{project_name}{A character vector containing the name for this new
project. Must be a valid directory name for your file system.}

\item{path}{A path where you want to store your new project.}
}
\description{
This function will create all of the scaffolding for a new ENR project.
It will set up all of the relevant directories and their initial
contents.
}
\examples{

\dontrun{project_create('new_enr_project', path = "/"}

}
