% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skelpear.R
\docType{package}
\name{skelpear-package}
\alias{skelpear}
\alias{skelpear-package}
\title{skelpear: Skelpear is a part of Pearsonverse}
\description{
Package for building standardised project environment that facilitates collaboration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pearsonplc/skelpear}
  \item Report bugs at \url{https://github.com/pearsonplc/skelpear/issues}
}

}
\keyword{internal}
