% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_pkg.R
\name{snapshot_pkg}
\alias{snapshot_pkg}
\title{Create a snapshot of your session info}
\usage{
snapshot_pkg()
}
\description{
A function which creates a file with your local session info (all packages you use) in config/packages.dcf file.
}
\examples{
\dontrun{
 snapshot_pkg()
 }
}
