% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skelpear.R
\docType{package}
\name{skelpear-package}
\alias{skelpear}
\alias{skelpear-package}
\title{skelpear: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bitbucket.pearson.com/projects/EF/repos/skelpear/browse}
}

}
\keyword{internal}
